moved redux action files to actions folder
